Added "dev" command for running the server in watch mode
Added the "dev" command to the package.json file, which allows the server to be run using Node.js in watch mode. This enables automatic server restarts when files are changed, streamlining the development process.

This feature is available starting from Node.js version 18.0.0, which introduced support for the `--watch` flag.